# backend/docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Spark Dating API
  version: 1.0.0
  description: API documentation for the Spark Dating backend
servers:
  - url: http://localhost:5000
    description: Development Server
  - url: http://localhost:5010
    description: RC1 Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /api/interactions/swipe:
    post:
      summary: Swipe on a user (LIKE, PASS, SUPER_LIKE)
      tags:
        - Interactions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                targetId:
                  type: string
                action:
                  type: string
                  enum: [LIKE, PASS, SUPER_LIKE]
      responses:
        '200':
          description: Swipe processed (asynchronously queued)
...
  /api/swipe:
    post:
      summary: Swipe on a user
      tags:
        - Swipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                targetId:
                  type: string
                rating:
                  type: string
                  enum: [like, nope]
      responses:
        '200':
          description: Swipe recorded
...
  /api/swipe/matches/{matchId}/messages:
    get:
      summary: Get messages for a match
      tags:
        - Swipe
      parameters:
        - in: path
          name: matchId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of messages
  /api/swipe/matches/{matchId}/read:
    post:
      summary: Mark messages as read
      tags:
        - Swipe
      parameters:
        - in: path
          name: matchId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Messages marked as read
...